feat(client): decompose DashboardGrid into composable pieces (#800)
DashboardGrid bundled its toolbar with no way to suppress or replace it,
forcing embedders to fork or fight internal CSS. Expose the seams instead:
the edit/save state machine was already extracted into a per-instance store
+ useDashboard() hook; only the JSX layer was monolithic.

Introduce a composable API under src/client/components/dashboard/:

  <DashboardProvider config={config} editable onSave={onSave}>
    <DashboardToolbar />      // or a host toolbar via useDashboardContext()
    <DashboardFilterBar />
    <DashboardGridSurface />
    <DashboardModals />
  </DashboardProvider>

- DashboardProvider owns the Zustand store; DashboardCoordinator runs
  useDashboard() + the responsive/scroll/drag machinery once and publishes
  everything through DashboardContext.
- The four pieces are pure context consumers; useDashboardContext() lets a
  host render its own toolbar while drizzle-cube owns the state machine.
- DashboardGrid is now a thin back-compat composition of the four pieces and
  gains a hideToolbar?: boolean prop (hides both the top bar and the floating
  toolbar).
- AnalyticsDashboard drops its now-redundant DashboardStoreProvider wrapper
  (DashboardProvider owns the store).
- Convert the bare English strings that moved (modal titles, edit-mode hint,
  delete-confirm modal) to existing t(...) keys; no locale additions.
- Export DashboardProvider/Toolbar/FilterBar/GridSurface/Modals and
  useDashboardContext from client and components entry points.

Adds tests/client/components/dashboard/DashboardProvider.test.tsx covering the
context guard, custom-toolbar control, hideToolbar, and composition parity.
