+++ src/client/components.ts
@@ -11,6 +11,16 @@ export { default as AnalyticsDashboard } from './components/AnalyticsDashboard'
 
 // Layout components
 export { default as DashboardGrid } from './components/DashboardGrid'
+// Composable dashboard pieces (DashboardGrid is a back-compat composition of these)
+export {
+  DashboardProvider,
+  DashboardToolbar,
+  DashboardFilterBar,
+  DashboardGridSurface,
+  DashboardModals,
+  useDashboardContext,
+} from './components/dashboard'
+export type { DashboardContextValue, DashboardProviderProps } from './components/dashboard'
 export { default as PortletContainer } from './components/PortletContainer'
 
 // Modals and configuration
